+++ desktop/frontend/src/pages/SearchPage.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import WorkspaceShell from '@/modules/workspace/components/WorkspaceShell.vue'
-</script>
-
-<template>
-  <WorkspaceShell context="search" />
-</template>